LayoutTests/http/tests/security/xssAuditor/object-tag-javascript-url-expected.txt

CONSOLE MESSAGE: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?test=/security/xssAuditor/object-tag-javascript-url.html&wait-for-load=1&q=%3Cobject%20id=%27object%27%20data=%27javascript:alert(document.domain)%27%3E%3C/object%3E' because its source code was found within the request. The auditor was enabled because the server did not send an 'X-XSS-Protection' header.
